# Fix Issue 17286 - A function for comparing two digests securely