Given two strings that represent two crypto digests, using std.algorithm.equal to compare them leaves your web application open to timing attacks because it has a short circuit, i.e. it returns `false` on the first inequality it finds.

The attack comes from allowing attacker to brute force your HMAC key for example. See this article for more information, evidence that the attack exists, and why Java gets it wrong: https://codahale.com/a-lesson-in-timing-attacks/

The solution is to have a string comparison that will always be constant time given two strings of the same length.

Ping @andralex, this is a name addition.