@JackStouffer 2017-03-29
Examples of this function is other languages: 

https://docs.python.org/2/library/hmac.html#hmac.compare_digest
http://www.rubydoc.info/github/rack/rack/master/Rack/Utils.secure_compare