@braddr 2017-03-29
Constant time algorithms are an important category, so I'm in favor of this in spirit, but the placement is questionable and I'd rather the name reflect the constant-ness rather than the secure-ness.