@quickfur 2017-03-29
I'm tempted to say this should go in `std.algorithm.comparison`, but then most people who need to know about this function may not see it when looking up digest-related Phobos modules.

But in any case, I think `ConstantTimeCompare` or something along those lines might be a better name for this.