@JackStouffer 2017-03-29
Added constraints. I think that the name is alright but I'm open to suggestions. I would like to know Andrei's opinion on the matter though.

I also don't know of any use for this function outside of digests, so that's why it's where it is. Let me know if there are other applications for this algorithm.