@JackStouffer 2017-03-30
> return lhs[i .. $] == rhs[i .. $] && lhs[0 .. i].retro == rhs[0 .. i].retro;

According to my benchmark, it's slower

```
$ dmd -O -inline -release test.d && ./test
old same	        538 ms and 300 μs
new same	1 sec, 162 ms, 32 μs, and 2 hnsecs
old different	545 ms, 930 μs, and 4 hnsecs
new different	696 ms, 248 μs, and 7 hnsecs
```

code

```d
import std.stdio;
import std.range;
import std.algorithm;
import std.traits;
import std.container;
import std.meta;
import std.conv;
import std.datetime;

enum testCount = 10_000_000;

bool secureCompare1(T)(T[] s1, T[] s2) if (isIntegral!T || isSomeChar!T)
{
    if (s1.length != s2.length)
        return false;

    int result = 0;
    foreach (i; 0 .. s1.length)
        result |= s1[i] ^ s2[i];

    return result == 0;
}

bool secureCompare2(T)(T[] lhs, T[] rhs) if (isIntegral!T || isSomeChar!T)
{
    import std.random : uniform;

    if (lhs.length != rhs.length)
        return false;

    immutable i = uniform(0, lhs.length);
    return lhs[i .. $] == rhs[i .. $] && lhs[0 .. i].retro == rhs[0 .. i].retro;
}

void main()
{
    auto hex1 = "02CA3484C375EDD3C0F08D3F50D119E610779018";
    auto hex2 = "02CA3484C375EDD3C0F08D3F50D119E610779018";
    auto hex3 = "A10BBD48B3BBD5765EF2F7CAED0D6BE40A156AE7";

    auto result = to!Duration(benchmark!(() => secureCompare1(hex1, hex2))(testCount)[0]);
    auto result2 = to!Duration(benchmark!(() => secureCompare2(hex1, hex2))(testCount)[0]);

    writeln("old same", "\t", result);
    writeln("new same", "\t", result2);

    auto result3 = to!Duration(benchmark!(() => secureCompare1(hex1, hex3))(testCount)[0]);
    auto result4 = to!Duration(benchmark!(() => secureCompare2(hex1, hex3))(testCount)[0]);

    writeln("old different", "\t", result3);
    writeln("new different", "\t", result4);
}
```