@andralex 2017-03-30
Yah no surprise it's slower for short inputs. I guess I'm fine assuming the prevalent inputs are SHA hashes. Speaking of which, we should make this work with ranges.