@JackStouffer 2017-03-30
>we should make this work with ranges

will do