@JackStouffer 2017-03-30
Updated the docs and made it work with ranges.