@quickfur 2017-03-31
@andralex  Your alternative will not thwart a determined attacker if the (P)RNG you use to choose `i` is not cryptographically secure.   Namely, using anything from `std.random` will not work.