@andralex 2017-04-01
@quickfur here's the code I used. It's 2x slower on the passing comparison and some 15% faster on the failing comparison:

```d
bool secureCompare2(T)(T[] lhs, T[] rhs) if (isIntegral!T || isSomeChar!T)
{
    import std.random;
    static Xorshift32 rnd;
    static bool init;
    if (!init)
    {
        rnd = typeof(rnd)(unpredictableSeed);
        init = true;
    }

    if (lhs.length != rhs.length)
        return false;

    immutable i = uniform(0, lhs.length, rnd);
    return lhs[i .. $] == rhs[i .. $] && lhs[0 .. i].retro == rhs[0 .. i].retro;
}
```

Now, assume `unpredictableSeed` is cryptographically secure. Would that make this algorithm cryptographically secure?