@JackStouffer 2017-04-01
I'm by no means an expert. I came across information on this issue when implementing HMAC signing of messages in my project and was surprised that D didn't have a solution.

Could you describe or link to some info on the copying issue? Won't this also be an issue in any scripting language?