@jpf91 2017-04-01
> Could you describe or link to some info on the copying issue?

I don't know any link directly covering this so I'll try to explain:

First we have to assume that we want to keep the hash secret. Private or symmetric encryption keys are often derived from hashes (HKDF, used in the noise protocol, ...).

Let's see how libsodium handles this: It has extra functions for [secure memory](https://download.libsodium.org/doc/helpers/memory_management.html):
* A special `memzero` function which avoids being removed by dead code elimination,
* a `mlock` function to prevent memory landing in swap or core dumps
* functions to prevent any access, read-only access or read-write access for a memory location
* a malloc which adds guard pages around allocated memory (see the sodium manual for details)

All these functions are used to prevent sensitive key material from leaking by some kind of buffer overflows or out of bound reads. Such memory 'leakage' is a much more common problem than situations where an attacker gains complete access to a server and could read keys from files. With certain protocols (e.g. noiseprotocol.org) you only need the private key in memory for a short time to derive a ephemeral key. After this you want your private key cleared from memory completely. All these functions also require that you know exactly where you keep your secret data in memory.

In libsodium you [explicitly manage the memory for hash state](https://download.libsodium.org/doc/hashing/generic_hashing.html) and hash result values so it's possible to use these secure memory functions.

So what's the problem with std.digest? The hash state is a struct, so it's pass by value. This can be fixed by a simple by-reference wrapper. But the `finish` methods can cause problems:

```d
ubyte[32] finish()
{
    ubyte[32] result = state;
    result[1] &= 42;
    return result;
}

void bar()
{
    auto secretKey = finish();
    // do something
    secureClearMemory(secretKey);
}
```

Without certain compiler optimization you might actually end up with two copies of the secret data: `result` might be in a different location on stack than `secretKey`. So the `secureClear` function will only clear one copy but the secret data will still be in stack memory. And depending on your function call patters and function stack sizes it might stay there for quite some time waiting for attackers scanning the applications memory....

Another problem is that you can't force the result value to be only in secure memory and never on the stack.

Now compiler optimizations, NRVO and the fact that large return values are usually passed by hidden reference (depends on the system ABI) usually help to avoid such copies. But I wouldn't want to rely on such optimizations for crypto code (Always remember that in certain countries in the worst case unsafe cryptography could be life-threatening).

Some related links:
https://www.securecoding.cert.org/confluence/display/cplusplus/MSC06-CPP.+Be+aware+of+compiler+optimization+when+dealing+with+sensitive+data
http://stackoverflow.com/questions/10683941/clearing-memory-securely-and-reallocations
http://www.cs.utsa.edu/~shxu/dsn07.pdf
http://stackoverflow.com/questions/1263350/cryptography-best-practices-for-keys-in-memory

And more related to the original pull request: A MSVC optimization which rendered a safe constant-time algorithm unsafe:
https://research.kudelskisecurity.com/2017/01/16/when-constant-time-source-may-not-save-you/

> Won't this also be an issue in any scripting language?

Probably, depends a lot on the language implementation.